[ir][refactor] First step to move Frontendd IR to its own file
