+++ taichi/ir/frontend.h
@@ -4,6 +4,7 @@
 #include "taichi/common/util.h"
 #include "taichi/common/dict.h"
 #include "taichi/util/io.h"
+#include "taichi/ir/frontend_ir.h"
 
 namespace taichi {
 static_assert(
