+++ taichi/transforms/lower_ast.cpp
@@ -1,6 +1,7 @@
-#include "taichi/ir/ir.h"
-
 #include <unordered_set>
+
+#include "taichi/ir/ir.h"
+#include "taichi/ir/frontend_ir.h"
 
 TLANG_NAMESPACE_BEGIN
